+++ web/static/js/phoenix.js
@@ -12,6 +12,8 @@
 // The `Socket` constructor takes the mount point of the socket
 // as well as options that can be found in the Socket docs,
 // such as configuring the `LongPoller` transport, and heartbeat.
+// Socket params can also be passed as an option for default, but
+// overridable channel params to apply to all channels.
 //
 //
 // ## Channels
@@ -305,6 +307,7 @@ export class Socket {
   // opts - Optional configuration
   //   transport - The Websocket Transport, ie WebSocket, Phoenix.LongPoller.
   //               Defaults to WebSocket with automatic LongPoller fallback.
+  //   params - The defaults for all channel params, ie `{user_id: userToken}`
   //   heartbeatIntervalMs - The millisec interval to send a heartbeat message
   //   reconnectAfterMs - The millisec interval to reconnect after connection loss
   //   logger - The optional function for specialized logging, ie:
@@ -326,6 +329,7 @@ export class Socket {
     this.logger               = opts.logger || function(){} // noop
     this.longpoller_timeout   = opts.longpoller_timeout || 20000
     this.endPoint             = this.expandEndpoint(endPoint)
+    this.params               = opts.params || {}
   }
 
   protocol(){ return location.protocol.match(/^https/) ? "wss" : "ws" }
@@ -417,8 +421,12 @@ export class Socket {
     this.channels = this.channels.filter( c => !c.isMember(chan.topic) )
   }
 
-  chan(topic, params){
-    let chan = new Channel(topic, params, this)
+  chan(topic, chanParams = {}){
+    let mergedParams = {}
+    for(var key in this.params){ mergedParams[key] = this.params[key] }
+    for(var key in chanParams){ mergedParams[key] = chanParams[key] }
+
+    let chan = new Channel(topic, mergedParams, this)
     this.channels.push(chan)
     return chan
   }