+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 * Enhancements
   * [Controller] Add `register_after_action/2` callback to be invoked after the controlle action has run.
 
+* Javascript client enhancements
+  * Add socket params to apply default, overridable params to all channel params.
+
 * Backward incompatible changes
   * [Controller] `plug :action` has been removed and the action is called automatically.